# Use reflect type as pcache key